Safari-611.1.10/LayoutTests/accessibility/aria-switch-checked-expected.txt

X
X
X
This tests that ARIA switches correctly handle the aria-checked attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS widget.isChecked is false
PASS widget.isChecked is true
PASS widget.isChecked is false
PASS widget.isChecked is true
PASS widget.isChecked is false
PASS successfullyParsed is true

TEST COMPLETE
